tokobarang/requirements.txt
asgiref==3.7.2
astroid==2.3.3
backports.zoneinfo==0.2.1
colorama==0.4.4
crispy-bootstrap4==2022.1
Django==4.2.8
django-crispy-forms==2.0
django-filter==23.5
django-login-required-middleware==0.9.0
django-random-id-model==0.1.1
django-widget-tweaks==1.4.5
gunicorn==21.2.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
packaging==23.2
pylint==2.4.4
pytz==2022.5
six==1.15.0
sqlparse==0.4.4
typing-extensions==4.9.0
wrapt==1.11.2
